+++ htdocs/api/v0.0.3-dev/candidates/Visit.php
@@ -92,11 +92,18 @@ class Visit extends \Loris\API\Candidates\Candidate
     public function handleGET()
     {
         $SubProjTitle = $this->Timepoint->getData("SubprojectTitle");
+        $centerID     = $this->Timepoint->getData("CenterID");
+        $center       = $this->DB->pselectRow(
+            "SELECT Name FROM psc WHERE CenterID =:cid",
+            array('cid' => $centerID)
+        );
+        $centerName   = $center['Name'];
 
         $this->JSON = [
                        "Meta" => [
                                   "CandID"  => $this->CandID,
                                   'Visit'   => $this->VisitLabel,
+                                  'Site'    => $centerName,
                                   'Battery' => $SubProjTitle,
                                  ],
                       ];
@@ -169,8 +176,8 @@ class Visit extends \Loris\API\Candidates\Candidate
             $this->safeExit(0);
 
         }
-        // This version of the API does not handle timepoint creation
-        // when users are at multiple sites
+        // When users are at multiple sites, the API requires
+        // siteName as an input to timepoint creation
         $user      = \User::singleton();
         $centerIDs = $user->getCenterIDs();
         $num_sites = count($centerIDs);
@@ -179,13 +186,42 @@ class Visit extends \Loris\API\Candidates\Candidate
             $this->error("You are not affiliated with any site");
             $this->safeExit(0);
         } else if ($num_sites > 1) {
-            $this->header("HTTP/1.1 501 Not Implemented");
-            $this->error(
-                "This API version does not support timepoint creation " .
-                "by users with multiple site affiliations. This will be ".
-                "implemented in a future release."
-            );
-            $this->safeExit(0);
+            if (!isset($this->ReceivedJSON['Meta']['Site'])
+            ) {
+                $this->header("HTTP/1.1 400 Bad Request");
+                $this->error(
+                    "Users affiliated with multiple sites " .
+                    "need to specify the name of the site at " .
+                    "which the visit took place."
+                );
+                $this->safeExit(0);
+            } else {
+                $siteName = $this->ReceivedJSON['Meta']['Site'];
+                foreach ($centerIDs as $key => $centerID) {
+                    $center = $this->DB->pselectRow(
+                        "SELECT CenterID as ID, Name FROM psc WHERE CenterID =:cid",
+                        array('cid' => $centerID)
+                    );
+                    $allUserSites[$centerID] = $center['Name'];
+                }
+                //Get the CenterID from the provided SiteName
+                $centerID = array_search($siteName, $allUserSites);
+                if (!in_array($centerID, $centerIDs)) {
+                    $this->header("HTTP/1.1 401 Unauthorized");
+                    $this->error(
+                        "You are creating a visit at a site you " .
+                        "are not affiliated with."
+                    );
+                    $this->safeExit(0);
+                }
+                $this->createNew(
+                    $this->CandID,
+                    $subprojectID,
+                    $this->VisitLabel,
+                    $centerID
+                );
+                $this->header("HTTP/1.1 201 Created");
+            }
         } else {
             $centerID          = $centerIDs[0];
             $candidateCenterID = \Candidate::singleton($this->CandID)
@@ -196,7 +232,12 @@ class Visit extends \Loris\API\Candidates\Candidate
                 $this->safeExit(0);
             }
             // need to extract subprojectID
-            $this->createNew($this->CandID, $subprojectID, $this->VisitLabel);
+            $this->createNew(
+                $this->CandID,
+                $subprojectID,
+                $this->VisitLabel,
+                $centerID
+            );
             $this->header("HTTP/1.1 201 Created");
         }
     }
@@ -211,13 +252,14 @@ class Visit extends \Loris\API\Candidates\Candidate
      * @param integer $subprojectID The subproject for the new visit
      * @param string  $VL           The visit label of the visit to
      *                              be created
+     * @param integer $CID          The ID of the center of the visit
      *
      * @return none
      */
-    function createNew($CandID, $subprojectID, $VL)
+    function createNew($CandID, $subprojectID, $VL, $CID)
     {
         try {
-            \TimePoint::isValidVisitLabel($CandID, $subprojectID, $VL);
+            \TimePoint::isValidVisitLabel($CandID, $subprojectID, $VL, $CID);
         } catch (\LorisException $e) {
             $this->header("HTTP/1.1 400 Bad Request");
             $this->error($e->getMessage());
