--- htdocs/api/v0.0.3/candidates/visits/images/format/Raw.php
@@ -1,95 +1,0 @@
-<?php
-/**
- * Handles API requests for the candidate's visit
- *
- * PHP Version 5.5+
- *
- * @category Main
- * @package  API
- * @author   Dave MacFarlane <david.macfarlane2@mcgill.ca>
- * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
- * @link     https://www.github.com/aces/Loris/
- */
-namespace Loris\API\Candidates\Candidate\Visit\Imaging\Format;
-require_once __DIR__ . '/../Image.php';
-/**
- * Handles API requests for the candidate's visit. Extends
- * Candidate so that the constructor will validate the candidate
- * portion of URL automatically.
- *
- * @category Main
- * @package  API
- * @author   Dave MacFarlane <david.macfarlane2@mcgill.ca>
- * @license  http://www.gnu.org/licenses/gpl-3.0.txt GPLv3
- * @link     https://www.github.com/aces/Loris/
- */
-class Raw extends \Loris\API\Candidates\Candidate\Visit\Imaging\Image
-{
-    /**
-     * Construct a visit class object to serialize candidate visits
-     *
-     * @param string $method     The method of the HTTP request
-     * @param string $CandID     The CandID to be serialized
-     * @param string $VisitLabel The visit label to be serialized
-     * @param string $Filename   The file name whose raw data we want
-     */
-    public function __construct($method, $CandID, $VisitLabel, $Filename)
-    {
-        $requestDelegationCascade = $this->AutoHandleRequestDelegation;
-
-        $this->AutoHandleRequestDelegation = false;
-
-        if (empty($this->AllowedMethods)) {
-            $this->AllowedMethods = ['GET'];
-        }
-
-        parent::__construct($method, $CandID, $VisitLabel, $Filename);
-
-        if ($requestDelegationCascade) {
-            $this->handleRequest();
-        }
-
-    }
-
-    /**
-     * Handles a GET request
-     *
-     * @return none, but populates $this->JSON
-     */
-    public function handleGET()
-    {
-        $mincfile = $this->getFullPath();
-        if (`which minctoraw`) {
-            $this->Header("Content-Type: application/x.raw");
-            passthru(
-                "minctoraw -byte -unsigned -normalize "
-                . escapeshellarg($mincfile)
-            );
-            $this->safeExit(0);
-        }
-
-        $this->header("HTTP/1.1 500 Internal Server Error", true, 500);
-        $this->error("Minc Tools not installed on server");
-    }
-
-
-    /**
-     * Calculate the entity tag for this URL
-     *
-     * @return string
-     */
-    public function calculateETag()
-    {
-        return null;
-    }
-}
-
-if (isset($_REQUEST['PrintRawFormat'])) {
-    $obj = new Raw(
-        $_SERVER['REQUEST_METHOD'],
-        $_REQUEST['CandID'],
-        $_REQUEST['VisitLabel'],
-        $_REQUEST['Filename']
-    );
-}
-?>
